Fix MSCP crash for older VMS
(Track,Group,Cylinder) = (1,1,1) causes some kind of overflow in older VMS (e.g. 5.5-2) on startup.

Using a fixed track size of 16 and limiting cylinders to 2047 resolves the issue. MSCP disks being limted to a max size of 4GB is not a problem; this is larger than the biggest DEC drive of the era, the 1.5GB RA92.